experiments/exp_requirements.txt
torch
scipy
numpy
scikit-image
tqdm
mpi4py
matplotlib
wget
pathlib
pillow
torchvision